Massive update of PD Manager application The current patien app is stripped from some features for increased stability
Clinical app may not run properly.

+++ src/main/java/com/pdmanager/core/buffers/AccDataBuffer.java
@@ -15,7 +15,7 @@ public class AccDataBuffer extends DataBuffer<AccData> {
     private String name;
 
 
-    public AccDataBuffer(long mBuffer, String pfolder, String pfileName) {
+    public AccDataBuffer(int mBuffer, String pfolder, String pfileName) {
 
 
         super(mBuffer, pfolder, pfileName);
